requirements.txt
azure_storage==0.37.0
rasterio==1.3.5.post1
rio_cogeo==3.5.0
